Share image font width fitting
